// src/types/floatButton.ts
export type FloatButtonShape = 'circle' | 'square';
export type FloatButtonSize = 'large' | 'small';
export type FloatButtonVariant = 'default' | 'primary' | 'success' | 'warning' | 'danger';
export type FloatButtonPosition = {
  top?: number;
  right?: number;
  bottom?: number;
  left?: number;
};